+++ src/renderers/stackingOffset.ts
@@ -47,6 +47,12 @@ export class StackingOffsetRenderer extends RendererBase {
                 throw new Error(`The requested board style (${ json.board.style }) is not supported by the '${ this.name }' renderer.`);
         }
 
+        // Rotate the board if requested
+        if (opts.rotate === 180) {
+            this.rotateBoard(draw);
+            gridPoints = gridPoints.map((r) => r.reverse()).reverse();
+        }
+
         // PIECES
         // Load all the pieces in the legend
         this.loadLegend(json, draw, opts);
@@ -100,6 +106,9 @@ export class StackingOffsetRenderer extends RendererBase {
                                     throw new Error(`The glyph you requested (${key}) does not contain the necessary information for scaling. Please use a different sheet or contact the administrator.`);
                                 }
                                 use.scale((this.cellsize / sheetCellSize) * 0.85);
+                                if (opts.boardClick !== undefined) {
+                                    use.click(() => opts.boardClick!(row, col, key));
+                                }
                             }
                         }
                     }